feat: Add toast notifications
Update the `TodoDetails` component to display the due date in a localized format using the `toLocaleString` method.

Modify the `TodoListItem` component to include a delete button that triggers the `deleteTodoHandler` function. This function sends a DELETE request to the server to delete the corresponding todo item. Additionally, update the due date display in the component to use the `toLocaleString` method for localization.