fix: Update TodoDetails to display the status of a todo
The code changes in TodoDetails.parts.jsx modify the component to display the status of a todo. It adds a new paragraph element to show the status, with a default value of "Not Started" if the status is undefined. This improvement enhances the user experience by providing visibility into the status of each todo.